fixes bug 1170727 - Removing Response class duplication in tests
